feat(microRestaurant) : implements article and menu

+++ microservice-restaurant/src/main/java/com/cesieats/microservicerestaurant/controller/RestaurantController.java
@@ -1,7 +1,9 @@
 package com.cesieats.microservicerestaurant.controller;
 
+import com.cesieats.microservicerestaurant.config.JwtUtil;
 import com.cesieats.microservicerestaurant.entity.Restaurant;
 import com.cesieats.microservicerestaurant.enums.Categorie;
+import com.cesieats.microservicerestaurant.error.AccesException;
 import com.cesieats.microservicerestaurant.error.RestaurantNotFoundException;
 import com.cesieats.microservicerestaurant.service.RestaurantService;
 import lombok.AllArgsConstructor;
@@ -15,15 +17,13 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 @RestController
 @RequestMapping("/restaurants")
+@AllArgsConstructor
 public class RestaurantController {
 
     @Autowired
     private final RestaurantService restaurantService;
 
-    public RestaurantController(RestaurantService restaurantService) {
-        this.restaurantService = restaurantService;
-    }
-
+    private final JwtUtil jwtUtil;
 
     @PostMapping("/create")
     @PreAuthorize("hasAuthority('RESTAURATEUR')")
@@ -34,14 +34,22 @@ public class RestaurantController {
 
     @PutMapping("update/{id}")
     @PreAuthorize("hasAuthority('RESTAURATEUR')")
-    public ResponseEntity<Restaurant> updateRestaurant(@PathVariable Long id, @Valid @RequestBody Restaurant restaurant) throws RestaurantNotFoundException {
+    public ResponseEntity<Restaurant> updateRestaurant(@RequestHeader("Authorization") String token, @PathVariable Long id, @Valid @RequestBody Restaurant restaurant) throws RestaurantNotFoundException {
+        String email = jwtUtil.extractEmail(token.substring(7));
+        if(!restaurantService.findRestaurantById(id).getCreatorEmail().equals(email)){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         Restaurant updated = restaurantService.updateRestaurant(id, restaurant);
         return ResponseEntity.ok(updated);
     }
 
     @DeleteMapping("/delete/{id}")
     @PreAuthorize("hasAuthority('RESTAURATEUR')")
-    public ResponseEntity<Void> deleteRestaurant(@PathVariable Long id) throws RestaurantNotFoundException {
+    public ResponseEntity<Void> deleteRestaurant(@RequestHeader("Authorization") String token, @PathVariable Long id) throws RestaurantNotFoundException {
+        String email = jwtUtil.extractEmail(token.substring(7));
+        if(!restaurantService.findRestaurantById(id).getCreatorEmail().equals(email)){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         restaurantService.deleteRestaurant(id);
         return ResponseEntity.noContent().build();
     }
@@ -50,6 +58,15 @@ public class RestaurantController {
         Restaurant restaurant = restaurantService.findRestaurantById(id);
         return ResponseEntity.ok(restaurant);
     }
+
+    @GetMapping("/myRestaurants")
+    @PreAuthorize("hasAuthority('RESTAURATEUR')")
+    public ResponseEntity<List<Restaurant>> getMyRestaurants(@RequestHeader("Authorization") String token) throws RestaurantNotFoundException {
+        String email = jwtUtil.extractEmail(token.substring(7));
+        List<Restaurant> restaurants = restaurantService.getMyRestaurants(email);
+        return ResponseEntity.ok(restaurants);
+    }
+
     @GetMapping("/all")
     public ResponseEntity<List<Restaurant>> getAllRestaurants() {
         List<Restaurant> restaurants = restaurantService.getAllRestaurant();
@@ -59,6 +76,11 @@ public class RestaurantController {
     @GetMapping("/categorie/{categorie}")
     public ResponseEntity<List<Restaurant>> getRestaurantsByCategorie(@PathVariable Categorie categorie) {
         List<Restaurant> restaurants = restaurantService.findByCategorie(categorie);
+        return ResponseEntity.ok(restaurants);
+    }
+    @GetMapping("/name/{name}")
+    public ResponseEntity<List<Restaurant>> getRestaurantsByName(@PathVariable String name) throws RestaurantNotFoundException {
+        List<Restaurant> restaurants = restaurantService.findRestaurantByName(name);
         return ResponseEntity.ok(restaurants);
     }
 }
