+++ microservice-restaurant/src/main/java/com/cesieats/microservicerestaurant/service/RestaurantServiceImpl.java
@@ -5,16 +5,21 @@ import com.cesieats.microservicerestaurant.entity.Restaurant;
 import com.cesieats.microservicerestaurant.error.RestaurantNotFoundException;
 import com.cesieats.microservicerestaurant.repository.RestaurantRepository;
 import lombok.AllArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.stream.Collectors;
+
 
 @Service
 @AllArgsConstructor
 public class RestaurantServiceImpl implements RestaurantService {
 
+
+
+    @Autowired
     private final RestaurantRepository restaurantRepository;
+
 
     @Override
     public List<Restaurant> getAllRestaurant() {
@@ -23,9 +28,7 @@ public class RestaurantServiceImpl implements RestaurantService {
 
     @Override
     public List<Restaurant> findByCategorie(Categorie categorie) {
-        return restaurantRepository.findAll().stream()
-                .filter(restaurant -> restaurant.getCategorie().equals(categorie))
-                .collect(Collectors.toList());
+        return restaurantRepository.findByCategorie(categorie);
     }
 
     @Override
@@ -42,8 +45,12 @@ public class RestaurantServiceImpl implements RestaurantService {
 
     @Override
     public void deleteRestaurant(Long id) throws RestaurantNotFoundException {
+        if(restaurantRepository.findById(id).isEmpty()){
+            throw new RestaurantNotFoundException("Restaurant non trouvé");
+        }
         restaurantRepository.deleteById(id);
     }
+
 
     @Override
     public Restaurant updateRestaurant(Long id, Restaurant restaurant) {
@@ -68,5 +75,19 @@ public class RestaurantServiceImpl implements RestaurantService {
         existing.setCategorie(restaurant.getCategorie());
         existing.setNbRate(restaurant.getNbRate());
         return restaurantRepository.save(existing);
+    }
+
+    @Override
+    public List<Restaurant> getMyRestaurants(String email) throws RestaurantNotFoundException {
+        return restaurantRepository.findByCreatorEmail(email);
+    }
+
+    @Override
+    public List<Restaurant> findRestaurantByName(String name) throws RestaurantNotFoundException {
+        List<Restaurant> restaurants = restaurantRepository.findByName(name);
+        if (restaurants.isEmpty()) {
+            throw new RestaurantNotFoundException("Aucun restaurant trouvé avec ce nom");
+        }
+        return restaurants;
     }
 }
